Move tokenInfo to app/lib
